feat: apply system of modules
