+++ .eslintrc.json
@@ -12,5 +12,6 @@
         "sourceType": "module"
     },
     "rules": {
+        "import/extensions": 0
     }
 }
