+++ src/sass/layouts/_keyboard.scss
@@ -51,7 +51,7 @@
 .on {
     background-color: $color-key-filler-on;
     border: 2px solid $color-key-boarder-light;
-  }
+}
 
 .key-backspace {
     min-width: 100px;
